requirements.txt
dj-database-url==0.5.0
Django==2.2.2
django-admin-actions==0.1.1
django-appconf==1.0.3
django-compressor==2.3
django-heroku==0.3.1
django-redis==4.10.0
django-sass-processor==0.7.3
gunicorn==19.9.0
libsass==0.19.1
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
rcssmin==1.0.6
redis==3.2.1
rjsmin==1.1.0
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
